Move digest related code to `digest` module
